% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_metadata_df}
\alias{create_metadata_df}
\title{Create metadata dataframe from BigWig filenames}
\usage{
create_metadata_df(map_id_vector = NULL, bw_files = NULL)
}
\arguments{
\item{map_id_vector}{A character vector of mapping IDs (optional).}

\item{bw_files}{A character vector of BigWig file paths (optional).}
}
\value{
A tibble containing metadata extracted from the filenames.
}
\description{
Create metadata dataframe from BigWig filenames
}
\examples{
\dontrun{
bw_files <- c(
 "/path/to/project1_batch1_H3K4me3_rerun_sample1_rep1.hg38.scaled.bw",
 "/path/to/project1_batch1_H3K9me3_rerun_sample2_pooled.hg38.unscaled.bw"
 )
 create_metadata_df(bw_files = bw_files)
 }
 #  # A tibble: 2 × 8
 #  bw_file                                      project_id batch   marker   rerun_id sample_id
 #  <chr>                                        <chr>      <chr>   <chr>    <chr>     <chr>
 #1 /path/to/project1_batch1_H3K4me3_rerun_sa… project1    batch1  H3K4me3  rerun     sample1
 #2 /path/to/project1_batch1_H3K9me3_rerun_sa… project1    batch1  H3K9me3  rerun     sample2
 #  replicate genome  scaling
 #  <chr>     <chr>   <chr>
 #1 rep1      hg38    scaled
 #2 pooled    hg38    unscaled

}
