% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.attach_msr_from_scaling_info}
\alias{.attach_msr_from_scaling_info}
\title{Discover and attach msr from scalinginfo file}
\usage{
.attach_msr_from_scaling_info(
  stats_summary,
  pipeline_output_path,
  scaling_info_file = NULL
)
}
\description{
Discover and attach msr from scalinginfo file
}
\keyword{internal}
