% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.discover_bigwig_files}
\alias{.discover_bigwig_files}
\title{Discover BigWig files in pipeline output directory}
\usage{
.discover_bigwig_files(path)
}
\description{
Discover BigWig files in pipeline output directory
}
\keyword{internal}
