% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.discover_stats_summary}
\alias{.discover_stats_summary}
\title{Discover stats_summary file in pipeline output directory}
\usage{
.discover_stats_summary(path)
}
\description{
Discover stats_summary file in pipeline output directory
}
\keyword{internal}
