% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.extract_marker_metadata}
\alias{.extract_marker_metadata}
\title{Extract marker names and metadata from BigWig filenames}
\usage{
.extract_marker_metadata(bw_files)
}
\description{
Extract marker names and metadata from BigWig filenames
}
\keyword{internal}
