% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.normalize_annotation_to_granges}
\alias{.normalize_annotation_to_granges}
\title{Normalize annotation input to GRanges}
\usage{
.normalize_annotation_to_granges(annotation, genome)
}
\description{
Normalize annotation input to GRanges
}
\keyword{internal}
