% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.resolve_stats_summary_input}
\alias{.resolve_stats_summary_input}
\title{Resolve QC input to a stats_summary-like data frame}
\usage{
.resolve_stats_summary_input(
  data = NULL,
  epk = NULL,
  stats_summary = NULL,
  fn_name = "plot_qc_stats"
)
}
\description{
Resolve QC input to a stats_summary-like data frame
}
\keyword{internal}
