% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epk.R
\name{.validate_bw_files_in_stats_summary}
\alias{.validate_bw_files_in_stats_summary}
\title{Validate that BigWig files match stats_summary$map_id}
\usage{
.validate_bw_files_in_stats_summary(bw_files, stats_summary)
}
\description{
Validate that BigWig files match stats_summary$map_id
}
\keyword{internal}
