% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpigenicR-package.R
\docType{package}
\name{EpigenicR-package}
\alias{EpigenicR}
\alias{EpigenicR-package}
\title{EpigenicR: Tools for Epigenomic Data Analysis and Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for processing, analyzing, and visualizing epigenomic data from ChIP-seq, methylation, and chromatin state experiments. Includes functions for quality control, metadata management, annotation retrieval, and creating comprehensive EPK (Epigenomic Package) objects for downstream analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epigenica/EpigenicR}
  \item Report bugs at \url{https://github.com/epigenica/EpigenicR/issues}
}

}
\author{
\strong{Maintainer}: Nima Rafati \email{nima.rafati@epigenica.org}

}
\keyword{internal}
