% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_marker_names}
\alias{extract_marker_names}
\title{Extract marker names from sample IDs}
\usage{
extract_marker_names(id, markers)
}
\arguments{
\item{id}{A character string representing the sample ID.}

\item{markers}{A vector of marker names to search for within the sample ID.}
}
\value{
A character vector containing the extracted marker names.
}
\description{
Extract marker names from sample IDs
}
\examples{
id <- "Sample1_H3K4me3_Rep1"
markers <- c("INPUT","H3K4me3","H3K9me3","H3K9ac","5mC","CXXC")
extract_marker_names(id, markers)

}
