% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap_cor_marker}
\alias{heatmap_cor_marker}
\title{Plot a correlation heatmap for one or more markers in an EPK experiment}
\usage{
heatmap_cor_marker(epk, exp_name, marker)
}
\arguments{
\item{epk}{An EPK object.}

\item{exp_name}{Name of the experiment (e.g. "protein_coding").}

\item{marker}{Name or vector of marker(s)/assay(s) (e.g. "H3K4me3" or c("5mC", "CXXC")).}
}
\value{
A ComplexHeatmap::Heatmap object (single marker) or a list of Heatmap objects (multiple markers).
}
\description{
Plot a correlation heatmap for one or more markers in an EPK experiment
}
\examples{
# Compute correlations first:
epk <- compute_all_cor(epk, exp_name = "genes")
# Plot heatmap for a single marker
heatmap_cor_marker(epk, exp_name = "genes", marker = "H3K4me3")
# Plot heatmaps for multiple markers
heatmap_cor_marker(epk, exp_name = "genes", marker = c("5mC", "CXXC"))
}
