% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_genes}
\alias{toy_genes}
\title{Toy gene coordinates}
\format{
A GRanges object with 5 ranges and metadata columns:
\describe{
  \item{seqnames}{Chromosome (chr22)}
  \item{ranges}{Genomic coordinates (start-end)}
  \item{strand}{Strand orientation (+/-)}
  \item{gene_name}{Gene symbol}
  \item{gene_id}{Ensembl gene identifier}
  \item{gene_type}{Gene biotype (protein_coding)}
}
}
\source{
Subset of human genome annotations (GENCODE)
}
\usage{
toy_genes
}
\description{
A small set of example gene coordinates for testing and examples.
Contains 5 protein-coding genes from chromosome 22.
}
\examples{
# Load toy gene coordinates
data(toy_genes)
toy_genes

# Assign to standard variable name to use with README examples
genes_coord_protein_coding <- toy_genes

# Use with wigglescout to process BigWig files
\dontrun{
toy_dir <- system.file("extdata", "toy_dataset", package = "EpigenicR")
bw_files <- list.files(toy_dir, pattern = "\\\\.bw$", full.names = TRUE)
bw_gr <- wigglescout::bw_loci(bw_files[1], loci = genes_coord_protein_coding)
}
}
\keyword{datasets}
